UI components - added Badge, Icon, Loader - Replaced BadgesContainer with common Container - Removed BadgesContainer - Moved Badge, Icon and Loader to ui components
