+++ src/components/CategoriesListItem/CategoriesListItemEditable.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { Input } from '../Controls';
 import { SidebarListItemControls, SidebarListItem } from '../SidebarBlocks';
-import Icon from '../Icon';
 import { useCategoryValidation } from '../../hooks';
 import { CategoryItem } from '../../models';
 import {
@@ -9,6 +8,7 @@ import {
   CategoriesListItemEditableDispatchToPropsMapResult,
 } from './CategoriesListItemEditableConnected';
 import { CategoriesOperationsActionTypes } from '../../action-types';
+import { Icon } from '../__ui__';
 
 interface CategoriesListItemEditableProps
   extends CategoriesListItemEditableStateToPropsMapResult,
