--- src/components/ContainerBlocks/BadgesContainer/BadgesContainer.scss
@@ -1,7 +1,0 @@
-.badges-container {
-  display: flex;
-
-  :not(:first-child) {
-    margin-left: 5px;
-  }
-}
