--- src/components/ContainerBlocks/BadgesContainer/BadgesContainer.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import './BadgesContainer.scss';
-
-const BadgesContainer: React.FC = props => {
-  return <div className="badges-container">{props?.children}</div>;
-};
-
-export default BadgesContainer;
