--- src/components/ContainerBlocks/BadgesContainer/index.ts
@@ -1,3 +1,0 @@
-import BadgesContainer from './BadgesContainer';
-
-export default BadgesContainer;
