+++ src/components/ContainerBlocks/index.ts
@@ -1,8 +1,7 @@
 import ContentWrapper from './ContentWrapper';
 import MainContainer from './MainContainer';
 import SectionContainer from './SectionContainer';
-import BadgesContainer from './BadgesContainer';
 import SectionTitle from './SectionTitle';
 import SectionPlaceholder from './SectionPlaceholder';
 
-export { ContentWrapper, MainContainer, SectionContainer, BadgesContainer, SectionTitle, SectionPlaceholder };
+export { ContentWrapper, MainContainer, SectionContainer, SectionTitle, SectionPlaceholder };
