+++ src/components/__ui__/Badge/Badge.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 
 .badge {
   font-size: 10px;
