+++ src/components/__ui__/Icon/Icon.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 
 .icon-container {
   display: flex;
