+++ src/components/__ui__/Loader/Loader.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 
 .loader-container {
   display: inline-flex;
