+++ src/components/__ui__/index.ts
@@ -1,3 +1,6 @@
 import Container from './Container';
+import Badge from './Badge';
+import Icon from './Icon';
+import Loader from './Loader';
 
-export { Container };
+export { Container, Badge, Icon, Loader };
